+++ src/renderer/dom/DomRendererRowFactory.test.ts
@@ -8,7 +8,7 @@ import { assert } from 'chai';
 import { DomRendererRowFactory } from './DomRendererRowFactory';
 import { DEFAULT_ATTR, NULL_CELL_CODE, NULL_CELL_WIDTH, NULL_CELL_CHAR } from '../../Buffer';
 import { FLAGS } from '../Types';
-import { BufferLine } from '../../TerminalLine';
+import { BufferLine } from '../../BufferLine';
 import { IBufferLine } from '../../Types';
 
 describe('DomRendererRowFactory', () => {
